clickable cards that redirect to corresponding task details page via task ID from database

+++ Final-Project-Restart-Your-Love/src/main/java/lv/restart/your/love/Final/Project/Restart/Your/Love/repository/TaskRepository.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Repository
 public interface TaskRepository extends CrudRepository<Task, Integer> {
 
-    Task findById(int id);
+    Task findById(long id);
     List<Task> findAll();
 
     //METHOD TO UPDATE task as done
